requirements.txt
discord.py
bs4
aiohttp
pymongo
jisha